Move logging module into its own file
No code changes, just relocating the module in preparation for the next
commit.
